Change edit campaign to file upload,remove list rule, and add description for campaign based on csv upload

+++ Modules/Campaign/Http/Controllers/ApiCampaign.php
@@ -120,17 +120,27 @@ class ApiCampaign extends Controller
 			));
 		}else{
 			$post = $request->json()->all();
+			if($request->has('import_file')&&!isset($post['conditions'])){
+				$post = $request->post();
+				$post=array_map(function($x){
+				return json_decode($x,true);},$post);
+				$post['conditions']=[];
+			}
+			if(!isset($post['campaign_description'])){
+				$post['campaign_description']='';
+			}
 		}
-		if(empty($post['conditions'])){
+		if(empty($post['conditions'])&&!$request->has('import_file')){
 			return [
 				'status'  => 'fail',
 				'messages'  => ['Rule must be filled in at least one']
 			];
 		}
 		$user = $request->user();
-		$data 						= [];
-		$data['campaign_title'] 	= $post['campaign_title'];
-		$data['id_user'] 			= $user['id'];
+		$data 							= [];
+		$data['campaign_title'] 		= $post['campaign_title'];
+		$data['campaign_description'] 	= $post['campaign_description'];
+		$data['id_user'] 				= $user['id'];
 
 		if(!empty($post['campaign_send_at'])){
 			$datetimearr 				= explode(' - ',$post['campaign_send_at']);
@@ -176,8 +186,8 @@ class ApiCampaign extends Controller
 
 		if($queryCampaign){
 			$data = [];
-
-			if(isset($post['id_campaign'])){
+			
+			if(isset($post['id_campaign'])&&!empty($post['conditions'])){
 				$deleteRuleParent = CampaignRuleParent::where('id_campaign','=',$post['id_campaign'])->get();
 				foreach ($deleteRuleParent as $key => $value) {
 					$value->rules()->delete();
@@ -187,24 +197,30 @@ class ApiCampaign extends Controller
 
 			if(isset($post['id_campaign'])) $data['id_campaign'] = $post['id_campaign'];
 				else $data['id_campaign'] = $queryCampaign->id_campaign;
-
-			$queryCampaignRule = MyHelper::insertCondition('campaign', $data['id_campaign'], $post['conditions']);
-			if(isset($queryCampaignRule['status']) && $queryCampaignRule['status'] == 'success'){
-				$resultrule = $queryCampaignRule['data'];
-			}else{
-				DB::rollBack();
+			if(!empty($post['conditions'])){
+				$queryCampaignRule = MyHelper::insertCondition('campaign', $data['id_campaign'], $post['conditions']);
+				if(isset($queryCampaignRule['status']) && $queryCampaignRule['status'] == 'success'){
+					$resultrule = $queryCampaignRule['data'];
+				}else{
+					DB::rollBack();
+					$result = [
+						'status'  => 'fail',
+						'messages'  => ['Create Campaign Failed']
+					];
+				}
 				$result = [
-					'status'  => 'fail',
-					'messages'  => ['Create Campaign Failed']
-				];
-			}
-
-			$result = [
 					'status'  => 'success',
 					'result'  => 'Set Campaign Information & Rule Success',
 					'campaign'  => $queryCampaign,
 					'rule'  => $resultrule
 				];
+			}else{
+				$result = [
+						'status'  => 'success',
+						'result'  => 'Set Campaign Information',
+						'campaign'  => $queryCampaign,
+					];
+			}
 		} else {
 			DB::rollBack();
 			$result = [
